@@ -1,10 +1,21 @@
 import { TestBed } from '@angular/core/testing';
 import { AppComponent } from './app.component';
+import { provideRouter } from '@angular/router';
+import { ThemeService } from './adapters/theme-service';
 
 describe('AppComponent', () => {
+  let mockThemeService: any;
   beforeEach(async () => {
+    mockThemeService = {
+      switchTheme: jasmine.createSpy('switchTheme')
+    };
+
     await TestBed.configureTestingModule({
       imports: [AppComponent],
+      providers: [
+        provideRouter([]),
+        { provide: ThemeService, useValue: mockThemeService }
+      ]
     }).compileComponents();
   });
 
@@ -14,16 +25,9 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have the 'my-app' title`, () => {
+  it(`should have the 'SGI' title`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('my-app');
-  });
-
-  it('should render title', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('h1')?.textContent).toContain('Hello, my-app');
+    expect(app.title).toEqual('SGI');
   });
 });
