@@ -1,16 +1,29 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { InventoryTableComponent } from './inventory-table.component';
+import { InventoryService } from '../../adapters/api/inventory.service';
+import { of } from 'rxjs';
+import { MessageService, ConfirmationService } from 'primeng/api';
 
 describe('InventoryTableComponent', () => {
   let component: InventoryTableComponent;
   let fixture: ComponentFixture<InventoryTableComponent>;
+  let mockInventoryService: any;
 
   beforeEach(async () => {
+    mockInventoryService = {
+      getProducts: jasmine.createSpy('getProducts').and.returnValue(of([])),
+      getProductsMini: jasmine.createSpy('getProductsMini').and.returnValue(of([]))
+    };
+
     await TestBed.configureTestingModule({
-      imports: [InventoryTableComponent]
+      imports: [InventoryTableComponent],
+      providers: [
+        { provide: InventoryService, useValue: mockInventoryService },
+        MessageService,
+        ConfirmationService
+      ]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(InventoryTableComponent);
     component = fixture.componentInstance;
